Move methods/classes used by all integration tests to helpers.py
Ensure setup_module is called at session level, making the test
suite not order dependent. Test files can be invoked independently.
